requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
Django==4.0.3
django-contrib-comments==2.2.0
django-extensions==3.1.5
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-simpleui==2022.3.19
django-threadedcomments==2.0
djangorestframework==3.13.1
idna==3.3
jsonfield==3.1.0
Markdown==3.3.4
martor==1.6.12
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
python-magic==0.4.25
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
swapper==1.3.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
